final-pjt-back/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.6.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
psycopg2==2.8.6